+++ draft-js-mention-plugin/src/MentionSuggestions/index.js
@@ -26,6 +26,12 @@ export default class MentionSuggestions extends Component {
     this.props.callbacks.onChange = this.onEditorStateChange;
   }
 
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.suggestions.size === 0 && this.state.isActive) {
+      this.closeDropdown();
+    }
+  }
+
   componentDidUpdate = (prevProps, prevState) => {
     if (this.refs.popover) {
       // In case the list shrinks there should be still an option focused.
@@ -45,6 +51,7 @@ export default class MentionSuggestions extends Component {
         prevState,
         props: this.props,
         state: this.state,
+        popover: this.refs.popover,
       });
       Object.keys(newStyles).forEach((key) => {
         this.refs.popover.style[key] = newStyles[key];
@@ -223,6 +230,10 @@ export default class MentionSuggestions extends Component {
     this.setState({
       isActive: true,
     });
+
+    if (this.props.onOpen) {
+      this.props.onOpen();
+    }
   };
 
   closeDropdown = () => {
@@ -239,11 +250,15 @@ export default class MentionSuggestions extends Component {
     this.setState({
       isActive: false,
     });
+
+    if (this.props.onClose) {
+      this.props.onClose();
+    }
   };
 
   render() {
     if (!this.state.isActive) {
-      return null;
+      return <noscript />;
     }
 
     const { theme = {} } = this.props;